LITE/code/requirements.txt
absl-py==2.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
equinox==0.11.4
etils==1.9.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
jax==0.4.31
jax-cuda12-pjrt==0.4.31
jax-cuda12-plugin==0.4.31
jaxlib==0.4.31
jaxopt==0.8.3
jaxtyping==0.2.31
trajax @ git+https://github.com/lenarttreven/trajax
Jinja2==3.1.4
kiwisolver==1.4.5
lineax==0.0.5
MarkupSafe==2.1.5
matplotlib==3.9.1
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.6.20
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optax==0.2.3
ott-jax==0.4.7
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.2
psutil==6.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
scipy==1.14.0
sentry-sdk==2.9.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.5
triton==2.3.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.2
wandb==0.17.4